requirements.development.txt
Django==1.8.2
Pillow==2.8.1
diff-match-patch==20121119
django-bootstrap-form==3.2
django-braces==1.8.0
django-debug-toolbar==1.3.0
django-debug-toolbar-template-timings==0.6.4
django-flat-theme==0.9.3
django-import-export==0.2.7
django-pipeline==1.5.0
django-simple-history==1.6.1
futures==3.0.2
ipython==3.1.0
jdcal==1.0
openpyxl==2.2.3
six==1.9.0
sqlparse==0.1.15
tablib==0.10.0
